First commit for the jumpLight project test_msa301.py tests the accelerometer connection and displays the acceleration along the x,y and z axes of the sensor.
test_shim.py tests the LED shim connection and makes a scrolling rainbow
display on the shim.

jumpLight.py connects to the accelerometer, detects a free-fall, measures
time till the end of free-fall and calculates the distance travelled during
free-fall. This distance is then used to light up a number of LEDs
proportional to the height jumped.

	new file:   jumpLight.py
	new file:   test_msa301.py
	new file:   test_shim.py